backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
stripe==7.7.0
razorpay==1.4.1
httpx==0.25.2
python-dotenv==1.0.0
email-validator==2.1.0
sentry-sdk[fastapi]==1.38.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
